parallelcut offset fix

+++ equipe45/src/main/java/Equipe45/domain/DTO/ParallelCutDTO.java
@@ -16,9 +16,9 @@ import java.util.UUID;
 public class ParallelCutDTO extends CutDTO {
 
     public UUID referenceID;
-    public float distance;
+    public int distance;
 
-    public ParallelCutDTO(UUID id, float depth, ToolDTO tool, UUID referenceID, float distance) {
+    public ParallelCutDTO(UUID id, float depth, ToolDTO tool, UUID referenceID, int distance) {
         super(id, depth, tool);
         this.referenceID = referenceID;
         this.distance = distance;
